# lib/public/swagger.yaml
openapi: 3.0.0
info:
  description: LowOnCloud API
  version: 1.0.0
  title: LowOnCloud API
  termsOfService: http://swagger.io/terms/
  contact:
    email: info@lowoncode.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
    
servers:
  # - url: https://lowoncode-cloud.herokuapp.com
  #   description: Production server (uses live data)
  - url: http://localhost:5678/_system
    description: Local server

security:
  - api_key: []
  
tags:
  - name: "components"
    description: ""
  - name: "instances"
    description: ""
  - name: "users"
    description: "Operations about user"
    
paths:
  /users:
    get:
      tags:
        - "users"
      description: ""
      responses:
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "405":
          description: Invalid input
          
  /components:
    get:
      tags:
        - "components"
      summary: Get all available components
      description: ""
      responses:
        "200":
          description: successful operation
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "405":
          description: Invalid input
      security:
        - api_key: []
    post:
      tags:
        - "components"
      summary: Create new instance
      description: ""
      responses:
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "405":
          description: Invalid input
  
    put:
      tags:
        - "components"
      summary: Update an existing pet
      description: ""
      responses:
        "200":
          description: successful operation
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instance"
        "400":
          description: Invalid status value
        "401":
          $ref: "#/components/responses/UnauthorizedError"

components:
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  securitySchemes:
    # api_key:
    #   type: apiKey
    #   name: X-API-Key
    #   in: header
    api_key:
      type: apiKey
      name: apikey
      in: header
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: integer
          format: int32
          description: User Status
    Instance:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nodes:
          type: integer
          format: int32
        shipDate:
          type: string
          format: date-time
        status:
          type: string
          description: Status
          enum:
            - placed
            - approved
            - delivered